putted sed in function

+++ dns.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
-#
+# TODO 
+# 	option to remove zones
 
-tld="drankenkas"
-sld="com"
+replace () {
+	sed -i "s/$1/$2/g" $3
+}
+
+tld="faka"
+sld="local"
 fqdn="$tld.$sld"
 
 echo " 
@@ -12,14 +17,17 @@ zone "FQDN" IN { type master;
 };
 " > /etc/named/template
 
-rm /etc/named/$fqdn
+if [ -e /etc/named/$fqdn ]; then
+	rm /etc/named/$fqdn
+fi
+
 cp /etc/named/template /etc/named/$fqdn
 
-sed -i "s/FQDN/$fqdn/g" "/etc/named/$fqdn"
-sed -i "s/TLD/$tld/g" "/etc/named/$fqdn"
+replace "FQDN" $fqdn "/etc/named/$fqdn"
+replace "TLD" $tld "/etc/named/$fqdn"
 
 echo "
 include \"/etc/named/$fqdn\";
 " >> /etc/named.conf
 
-sudo systemctl restart named.service
+#sudo systemctl restart named.service
